refactor: no switch. rent appium context
