+++ packages/typescript-private/device-server/src/device-webdriver/appium.device-webdriver.handler.ts
@@ -1,14 +1,14 @@
 import { Code, ErrorResultError, Platform, Serial } from '@dogu-private/types';
-import { errorify, HeaderRecord } from '@dogu-tech/common';
+import { HeaderRecord } from '@dogu-tech/common';
 import { RelayRequest, RelayResponse, SessionDeletedParam, WebDriverEndPoint } from '@dogu-tech/device-client-common';
-import { AppiumContextProxy } from '../appium/appium.context';
-import { AppiumRemoteContext } from '../appium/appium.remote.context';
+import { AppiumContextProxy, AppiumRemoteContextRental } from '../appium/appium.context.proxy';
 import { HttpRequestRelayService } from '../http-request-relay/http-request-relay.common';
 import { DoguLogger } from '../logger/logger';
 import { DeviceWebDriverHandler } from './device-webdriver.common';
 import { AppiumEndpointHandlerService } from './endpoint-handler/appium.service';
 
 export class AppiumDeviceWebDriverHandler implements DeviceWebDriverHandler {
+  private remoteRental: AppiumRemoteContextRental | null = null;
   constructor(
     private readonly platform: Platform,
     private readonly serial: Serial,
@@ -19,9 +19,7 @@ export class AppiumDeviceWebDriverHandler implements DeviceWebDriverHandler {
   ) {}
 
   async onRelayHttp(headers: HeaderRecord, request: RelayRequest): Promise<RelayResponse> {
-    if (this.appiumContextProxy.key !== 'remote') {
-      await this.appiumContextProxy.switchAppiumContext('remote', 'remote-webdriver-start');
-    }
+    this.remoteRental = await this.appiumContextProxy.rentRemote('AppiumDeviceWebDriverHandler');
 
     const httpRequestRelayHandler = this.httpRequestRelayService.getHandler(request.method);
     if (!httpRequestRelayHandler) {
@@ -32,13 +30,7 @@ export class AppiumDeviceWebDriverHandler implements DeviceWebDriverHandler {
       });
     }
 
-    const appiumRemoteContext = this.appiumContextProxy.getImpl(AppiumRemoteContext);
-    if (!appiumRemoteContext) {
-      throw new ErrorResultError(Code.CODE_UNEXPECTED_ERROR, 'appiumRemoteContext is null', {
-        platform: this.platform,
-        serial: this.serial,
-      });
-    }
+    const appiumRemoteContext = this.remoteRental.context;
 
     const endpoint = await WebDriverEndPoint.fromRelayRequest(request);
     const endpointHandler = this.appiumEndpointHandlerService.getHandler(endpoint.info.type);
@@ -59,15 +51,10 @@ export class AppiumDeviceWebDriverHandler implements DeviceWebDriverHandler {
   }
 
   async onSessionDeleted(headers: HeaderRecord, param: SessionDeletedParam): Promise<void> {
-    if (this.appiumContextProxy.key === 'remote') {
-      const appiumRemoteContext = this.appiumContextProxy.getImpl(AppiumRemoteContext);
-      if (appiumRemoteContext.sessionId === param.sessionId) {
-        try {
-          await this.appiumContextProxy.switchAppiumContext('builtin', 'remote-webdriver-done');
-        } catch (error) {
-          this.logger.error('Error while switching to builtin context', { error: errorify(error) });
-        }
-      }
+    if (!this.remoteRental) {
+      return;
     }
+    await this.remoteRental.release();
+    this.remoteRental = null;
   }
 }
